Organização das rotas para incluir login de usuário

+++ src/app/controllers/livro/DeleteLivroController.js
@@ -9,6 +9,10 @@ class DeleteLivroController{
     async delete(request, response) {
         const { id } = request.params;
 
+        if (!id) {
+            return response.status(404).json({Mensagem: "Livro não localizado."})
+        };
+
         const deletado = await this.service.delete(id);
 
         return response.status(200).json(deletado);
